requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
httpx==0.25.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# PDF and Resume Parsing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
camelot-py[cv]==0.10.1
pandas==2.1.3

# Google Services
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai==0.3.2

# Firebase
firebase-admin==6.2.0

# AI/ML Integration
openai==1.3.7

# Environment Variables
python-dotenv==1.0.0

# Additional dependencies for production
gunicorn==21.2.0